renamed search component
