+++ frontend/app/search/page.tsx
@@ -1,4 +1,4 @@
-import Header from "@/_ui/components/Header/Header";
+import Header from "@/_ui/components/Search/Search";
 export default function SearchPage() {
     return (
       <main>
